Updating, modernizing
Removed necessary code to look for RubyGems. 
Remove necessary 'service' prefix - if you're creating a service, you know what to call it.
Adding AppDirectory to parameters - this allows you to specify the directory context in which your code will run from (Useful for things like installing Rails as a service).